Load Markdown into Editor.js
This adds a method that will accept Markdown input and load it into the
Editor.js instance as blocks.

Internally, this performs a couple of steps:

1. Convert Markdown to HTML
2. Load HTML into Editor.js

By loading HTML into the Editor.js instance, we have the benefit of
being able to use the existing 'paste' functionality of Editor.js. This
is functionally the same as pasting HTML into the editor, where HTML
elements are converted into Blocks based on their type (e.g. H2 becomes
a Heading).