Add image tool with Netlify Function
I've added the Image tool for EditorJS - @editorjs/image

This requires a server-side API that will accept image file uploads and
respond with a URL so it can be rendered in the editor.

For the purposes of the demo, I've used Netlify Functions (serverless)
to accept file uploads and return them as Base64 encoded data URLs.
Uploaded images aren't stored anywhere, they're simply returned back to
the client as a Base64 string for displaying in the demo editor.